Accurate TM->TMW and TMW->TM mappings are now available.  I've verified this extensively and have full confidence that these mappings agree with Tony Duff's Tibetan! 5.1 documentation (except as described below).
To get them, I had to disregard Tony Duff's tables for a few glyphs: the
characters with ordinal 32 and 45 (space and hyphen in Roman ASCII,
space and tsheg in Tibetan).  For these glyphs, we must have mappings
from TibetanMachineSkt4.32 to something, etc., and those mappings were
not present.  I've normalized the mapping for these glyphs, as it is arbitrary
because the same two glyphs just appear fifteen times each.